pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openscad-mcp"
version = "0.1.0"
description = "MCP server for OpenSCAD 3D modelling via CLI"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
where = ["src"]
